optimization and cmake tweaks

+++ src/util.cpp
@@ -54,7 +54,7 @@ inline std::string colorize ( const std::string& text, Color color ) {
 inline std::string humanReadableSize ( std::ifstream::pos_type size ) {
 	const char* units[] = {"B", "KB", "MB", "GB", "TB"};
 	auto sizeDouble = static_cast<double>(size);
-	int unitIndex = 0;
+	size_t unitIndex = 0;
 
 	// Calculate the appropriate unit
 	while ( sizeDouble >= 1000.0 && unitIndex < std::size(units) - 1 ) {
@@ -70,7 +70,7 @@ inline std::string humanReadableSize ( std::ifstream::pos_type size ) {
 
 inline std::string humanReadableSpeed ( double speed ) {
 	const char* units[] = {"B/s", "KB/s", "MB/s", "GB/s", "TB/s"};
-	int unitIndex = 0;
+	size_t unitIndex = 0;
 
 	// Calculate the appropriate unit
 	while ( speed >= 1000.0 && unitIndex < std::size(units) - 1 ) {
